Using the footer, header, and hero from the components libraries and connecting the new sass styles.
