+++ package.json
@@ -5,7 +5,8 @@
   "homepage": "/data-catalog-frontend/build",
   "dependencies": {
     "axios": "^0.18.0",
-    "interra-data-catalog-components": "0.0.3",
+    "interra-data-catalog-components": "0.0.4",
+    "node-sass": "^4.11.0",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-router-dom": "^4.3.1",
