--- src/styles.css
@@ -1,17 +1,0 @@
-html {
-  /*Convert font size to base 10 for easier calculations (1rem = 10px)*/
-  font-size: 62.5%;
-}
-body {
-  font-size: 1.6rem;
-  font-family: 'Rubik', 'Open Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
-}
-h1 {
-  font-size: 3.6rem;
-}
-h2 {
-  font-size: 2.4rem;
-}
-h3 {
-  font-size: 1.6rem;
-}
