--- src/theme.js
@@ -1,19 +1,0 @@
-const defaultTheme = {
-  backgroundColor: '#f9fafb',
-  borderColor: '#A4B9C6',
-  primaryDark: '#003366',   // blueDark.
-  primaryMedium: '#0A77BD', // blueMedium.
-  primaryLight: '#27AAE1',  // blueLight.
-  primaryDust: '#F4F9FB',   // blueDust.
-  grayDark: '#323A45',
-  grayMedium: '#A7AAAC',
-  grayLight: '#dddddd',
-  grayDust: '#eeeeee',
-  accentDark: '#dfb100',   // yellowDark.
-  accentMedium: '#f2d13c', // yellowMedium.
-  accentLight: '#fff8e4',  // yellowLight.
-  headerColor: '#003366',
-  linkColor: '#003366',
-}
-
-export default defaultTheme;
