+++ src/theme/css/_button.scss
@@ -1,0 +1,37 @@
+// Button styles
+
+.btn {
+  box-shadow: 0 4px 7px rgba(0,0,0,0.2);
+  background-color: #FFD666;
+  color: #043C60;
+  border-radius: 100px;
+  touch-action: manipulation;
+  cursor: pointer;
+  white-space: nowrap;
+  display: inline-block;
+  font-weight: 600;
+  padding: 15px 40px;
+  margin: 1rem;
+  text-align: center;
+  letter-spacing: 1px;
+  text-decoration: none;
+  text-shadow: none;
+  text-align: center;
+  vertical-align: middle;
+}
+.btn:first-of-type {
+  margin-left: 0;
+  margin-right: 0;
+}
+.btn-hero {
+  background-color: #FFD666;
+  color: #043C60;
+  border: none;
+  text-transform: uppercase;
+  &:hover,
+  &:focus {
+    background: white;
+    transform: translateY(-3px);
+    box-shadow: 0 7px 15px rgba(0,0,0,0.5);
+  }
+}
