+++ src/theme/css/_variables.scss
@@ -1,0 +1,19 @@
+// Variables
+
+$backgroundColor: '#f9fafb';
+$borderColor: '#A4B9C6';
+$primaryDark: '#003366';   // blueDark.
+$primaryMedium: '#0A77BD'; // blueMedium.
+$primaryLight: '#27AAE1';  // blueLight.
+$primaryDust: '#F4F9FB';   // blueDust.
+$grayDark: '#323A45';
+$grayMedium: '#A7AAAC';
+$grayLight: '#dddddd';
+$grayDust: '#eeeeee';
+$accentDark: '#dfb100';   // yellowDark.
+$accentMedium: '#f2d13c'; // yellowMedium.
+$accentLight: '#fff8e4';  // yellowLight.
+$headerColor: '#003366';
+$textColor: '#3B3B3B';
+$linkColor: '#003366';
+$fontFamily: '"Rubik", "Open Sans", "Helvetica Neue", Helvetica, Arial, sans-serif;'
