+++ src/theme/default.js
@@ -1,0 +1,35 @@
+const defaultTheme = {
+  backgroundColor: '#f9fafb',
+  borderColor: '#A4B9C6',
+  primaryDark: '#003366',   // blueDark.
+  primaryMedium: '#0A77BD', // blueMedium.
+  primaryLight: '#27AAE1',  // blueLight.
+  primaryDust: '#F4F9FB',   // blueDust.
+  grayDark: '#323A45',
+  grayMedium: '#A7AAAC',
+  grayLight: '#dddddd',
+  grayDust: '#eeeeee',
+  accentDark: '#dfb100',   // yellowDark.
+  accentMedium: '#f2d13c', // yellowMedium.
+  accentLight: '#fff8e4',  // yellowLight.
+  headingColor: '#003366',
+  textColor: '#3B3B3B',
+  linkColor: '#003366',
+  fontFamily: '"Rubik", "Open Sans", "Helvetica Neue", Helvetica, Arial, sans-serif;',
+  navBarBackgroundColor: "#007BBC",
+  navBarBackgroundImage: "linear-gradient(to right, #007BBC 0%, #133F75 100%)",
+  navBarLink: "#eaeaea",
+  navBarLinkActive: "#FFFFFF",
+  navBarLinkHoverBack: "#002245",
+  navBarLinkActiveBack: "#305e92",
+  navBarLinkActiveHoverBack: "#f2d13c",
+  navBarLinkActiveHover: "#FFFFFF",
+  heroBackgroundColor: "#0A77BD",
+  heroTextColor: "#FFFFFF",
+  heroHeadingColor: "#FFFFFF",
+  footerBackgroundColor: '#004466',
+  footerText: '#FFFFFF',
+  footerLink: '#FFFFFF'
+}
+
+export default defaultTheme;
