+++ src/theme/styles.scss
@@ -1,0 +1,11 @@
+@vars {
+  src: url('css/_variables.scss');
+}
+@base {
+  src: url('css/_base.scss');
+}
+@button {
+  src: url('css/_button.scss');
+}
+
+
